+++ app/src/main/java/com/concordium/wallet/ui/more/export/ExportAccountKeysViewModel.kt
@@ -18,6 +18,7 @@ import com.concordium.wallet.data.model.ExportAccountKeys
 import com.concordium.wallet.data.model.Value
 import com.concordium.wallet.data.room.Account
 import com.concordium.wallet.data.util.FileUtil
+import com.concordium.wallet.util.Log
 import com.google.gson.Gson
 import kotlinx.coroutines.launch
 
@@ -28,10 +29,12 @@ class ExportAccountKeysViewModel(application: Application) : AndroidViewModel(ap
     val textResourceInt: MutableLiveData<Int> by lazy { MutableLiveData<Int>() }
     val accountData: MutableLiveData<AccountData> by lazy { MutableLiveData<AccountData>() }
 
-    fun saveFileToLocalFolder(destinationUri: Uri) {
-        viewModelScope.launch {
+    fun saveFileToLocalFolder(destinationUri: Uri) = viewModelScope.launch {
+        try {
             val accountKeys = AccountKeys(accountDataKeys, account.credential?.getThreshold() ?: 1)
-            val credentials = Credentials(account.credential?.getCredId() ?: "")
+            val credentials = Credentials(checkNotNull(account.credential?.getCredId()) {
+                "The credential must have an ID"
+            })
             val value = Value(accountKeys, credentials, account.address)
             val fileContent = Gson().toJson(
                 ExportAccountKeys(
@@ -43,6 +46,10 @@ class ExportAccountKeysViewModel(application: Application) : AndroidViewModel(ap
             )
             FileUtil.writeFile(destinationUri, "${account.address}.export", fileContent)
             textResourceInt.postValue(R.string.export_account_keys_file_exported)
+        } catch (e: Exception) {
+            Log.e("File export failed", e)
+
+            textResourceInt.postValue(R.string.app_error_general)
         }
     }