+++ app/src/main/java/com/concordium/wallet/ui/walletconnect/WalletConnectViewModel.kt
@@ -4,21 +4,41 @@ import android.app.Application
 import android.text.TextUtils
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
+import com.concordium.sdk.crypto.bulletproof.BulletproofGenerators
+import com.concordium.sdk.crypto.pedersencommitment.PedersenCommitmentKey
+import com.concordium.sdk.crypto.wallet.Network
+import com.concordium.sdk.crypto.wallet.web3Id.AcceptableRequest
+import com.concordium.sdk.crypto.wallet.web3Id.AccountCommitmentInput
+import com.concordium.sdk.crypto.wallet.web3Id.Statement.IdentityQualifier
+import com.concordium.sdk.crypto.wallet.web3Id.Statement.StatementType
+import com.concordium.sdk.crypto.wallet.web3Id.Statement.UnqualifiedRequestStatement
+import com.concordium.sdk.crypto.wallet.web3Id.UnqualifiedRequest
+import com.concordium.sdk.crypto.wallet.web3Id.Web3IdProof
+import com.concordium.sdk.crypto.wallet.web3Id.Web3IdProofInput
+import com.concordium.sdk.responses.accountinfo.credential.AttributeType
+import com.concordium.sdk.responses.cryptographicparameters.CryptographicParameters
+import com.concordium.sdk.transactions.CredentialRegistrationId
+import com.concordium.sdk.types.UInt32
 import com.concordium.wallet.App
 import com.concordium.wallet.BuildConfig
 import com.concordium.wallet.R
 import com.concordium.wallet.core.backend.BackendRequest
 import com.concordium.wallet.data.AccountRepository
+import com.concordium.wallet.data.IdentityRepository
 import com.concordium.wallet.data.backend.repository.ProxyRepository
+import com.concordium.wallet.data.cryptolib.AttributeRandomness
 import com.concordium.wallet.data.cryptolib.CreateAccountTransactionInput
 import com.concordium.wallet.data.cryptolib.CreateTransferOutput
 import com.concordium.wallet.data.cryptolib.SignMessageInput
 import com.concordium.wallet.data.cryptolib.StorageAccountData
 import com.concordium.wallet.data.model.AccountData
 import com.concordium.wallet.data.model.AccountNonce
+import com.concordium.wallet.data.model.GlobalParams
 import com.concordium.wallet.data.model.TransactionType
 import com.concordium.wallet.data.model.TransferCost
+import com.concordium.wallet.data.preferences.AuthPreferences
 import com.concordium.wallet.data.room.Account
+import com.concordium.wallet.data.room.Identity
 import com.concordium.wallet.data.room.WalletDatabase
 import com.concordium.wallet.data.walletconnect.AccountTransactionParams
 import com.concordium.wallet.data.walletconnect.AccountTransactionPayload
@@ -103,12 +123,20 @@ private constructor(
     private val allowedRequestMethods = setOf(
         REQUEST_METHOD_SIGN_AND_SEND_TRANSACTION,
         REQUEST_METHOD_SIGN_MESSAGE,
+        REQUEST_METHOD_VERIFIABLE_PRESENTATION,
     )
 
     private val accountRepository: AccountRepository by lazy {
         AccountRepository(WalletDatabase.getDatabase(getApplication()).accountDao())
     }
     private val proxyRepository: ProxyRepository by lazy(::ProxyRepository)
+
+    private val identityRepository: IdentityRepository by lazy {
+        IdentityRepository(WalletDatabase.getDatabase(getApplication()).identityDao())
+    }
+
+    private val authPreferences: AuthPreferences
+
 
     private lateinit var sessionProposalPublicKey: String
     private lateinit var sessionProposalNamespaceKey: String
@@ -124,8 +152,17 @@ private constructor(
     private lateinit var sessionRequestTransactionCost: BigInteger
     private lateinit var sessionRequestTransactionNonce: AccountNonce
     private lateinit var sessionRequestMessageToSign: String
-
+    private lateinit var sessionRequestIdentityRequest: UnqualifiedRequest
+    private lateinit var sessionRequestIdentityProofAccounts: MutableList<Account>
+    private lateinit var sessionRequestIdentityProofIdentities: MutableMap<Int, Identity>
+    private lateinit var sessionRequestIdentityProofProvable: ProofProvableState
     private val handledRequests = mutableSetOf<Long>()
+
+    enum class ProofProvableState {
+        Provable,
+        UnProvable,
+        NoCompatibleIssuer
+    }
 
     private val mutableStateFlow = MutableStateFlow<State>(State.Idle)
     val stateFlow: Flow<State> = mutableStateFlow
@@ -156,6 +193,8 @@ private constructor(
                         "\nevent=$event"
             )
         }
+
+        authPreferences = AuthPreferences(application)
     }
 
     fun initialize() {
@@ -441,6 +480,7 @@ private constructor(
                     selectedAccount = reviewState.selectedAccount,
                     accounts = accounts,
                     appMetadata = reviewState.appMetadata,
+                    identityProofPosition = null
                 )
             )
         }
@@ -456,11 +496,64 @@ private constructor(
                     "\nnewSelectedAccountId=${selectedAccount.id}"
         )
 
-        mutableStateFlow.tryEmit(
-            State.SessionProposalReview(
-                selectedAccount = selectedAccount,
-                appMetadata = selectionState.appMetadata
+        if (!selectionState.forIdentityProof) {
+            mutableStateFlow.tryEmit(
+                State.SessionProposalReview(
+                    selectedAccount = selectedAccount,
+                    appMetadata = selectionState.appMetadata
+                )
             )
+        } else {
+            onAccountSelectedIdentityProof(selectedAccount, selectionState.identityProofPosition!!)
+        }
+    }
+
+    fun onChooseAccountIdentityProof(position: Int) {
+        val proofRequestReview =
+            checkNotNull(state as? State.SessionRequestReview.IdentityProofRequestReview) {
+                "Choose account button can only be clicked in the proof request review state"
+            }
+
+        viewModelScope.launch {
+            val accounts = getAvailableAccounts()
+            val statement = sessionRequestIdentityRequest.credentialStatements[position]
+
+            Log.d(
+                "switching_to_account_selection:" +
+                        "\navailableAccounts=${accounts.size}"
+            )
+
+            val validAccounts = accounts.filter {
+                val identity = getIdentity(it) ?: return@filter false
+                isValidIdentityForStatement(identity, statement)
+            }
+
+            mutableStateFlow.emit(
+                State.AccountSelection(
+                    selectedAccount = proofRequestReview.chosenAccounts[position],
+                    accounts = validAccounts.ifEmpty { accounts },
+                    appMetadata = proofRequestReview.appMetadata,
+                    identityProofPosition = position
+                )
+            )
+        }
+    }
+
+    private fun createIdentityProofRequestState(currentStatement: Int): State.SessionRequestReview {
+        return State.SessionRequestReview.IdentityProofRequestReview(
+            connectedAccount = sessionRequestAccount,
+            appMetadata = sessionRequestAppMetadata,
+            request = sessionRequestIdentityRequest,
+            chosenAccounts = sessionRequestIdentityProofAccounts,
+            currentStatement = currentStatement,
+            provable = sessionRequestIdentityProofProvable
+        )
+    }
+
+    private fun onAccountSelectedIdentityProof(selectedAccount: Account, position: Int) {
+        sessionRequestIdentityProofAccounts[position] = selectedAccount
+        mutableStateFlow.tryEmit(
+            createIdentityProofRequestState(position)
         )
     }
 
@@ -592,6 +685,114 @@ private constructor(
         }
     }
 
+    private fun onInvalidRequest(responseMessage: String, e: Exception? = null) {
+        if (e == null) Log.e(responseMessage) else Log.e(responseMessage, e)
+        mutableEventsFlow.tryEmit(
+            Event.ShowFloatingError(
+                Error.InvalidRequest
+            )
+        )
+        respondError(responseMessage)
+        mutableStateFlow.tryEmit(State.Idle)
+    }
+
+    private fun isValidIdentityForStatement(
+        identity: Identity,
+        statement: UnqualifiedRequestStatement
+    ): Boolean {
+        return statement.idQualifier is IdentityQualifier
+                && (statement.idQualifier as IdentityQualifier).issuers.contains(identity.identityProviderId.toLong())
+                && statement.canBeProvedBy(getIdentityObject(identity))
+    }
+
+    /// Initializes session variables for proof request session.
+    /// Returns a boolean indicating whether initialization succeeded.
+    private suspend fun initializeProofRequestSession(params: String): Boolean {
+        if (!authPreferences.hasEncryptedSeed()) {
+            Log.d("A non-seed phrase wallet does not support identity proofs")
+            mutableEventsFlow.tryEmit(
+                Event.ShowFloatingError(
+                    Error.NotSeedPhraseWalletError
+                )
+            )
+            return false
+        }
+        try {
+            val wrappedParams = Gson().fromJson(params, WalletConnectParamsWrapper::class.java)
+            val identityProofRequest = UnqualifiedRequest.fromJson(wrappedParams.paramsJson)
+            sessionRequestIdentityRequest = identityProofRequest
+        } catch (e: Exception) {
+            onInvalidRequest("Failed to parse identity proof request parameters: $params", e)
+            return false
+        }
+
+        if (sessionRequestIdentityRequest.credentialStatements.any { it.statementType == StatementType.SmartContract }) {
+            onInvalidRequest("Web3Id statements are not supported.")
+            return false
+        }
+
+        try {
+            // Check that the statement is acceptable. Using the session account here to qualify
+            // is sufficient because we don't check that the statement is provable.
+            val network =
+                if (BuildConfig.ENV_NAME == "production") Network.MAINNET else Network.TESTNET
+            val credId =
+                CredentialRegistrationId.from(sessionRequestAccount.credential!!.getCredId()!!)
+            AcceptableRequest.acceptableRequest(sessionRequestIdentityRequest.qualify { stat ->
+                stat.qualify(
+                    credId,
+                    network
+                )
+            })
+        } catch (e: Exception) {
+            onInvalidRequest("Request received is not acceptable", e)
+            return false
+        }
+
+        val identities = HashMap<Int, Identity>()
+        identityRepository.getAllDone().forEach { identities[it.id] = it }
+        sessionRequestIdentityProofIdentities = identities
+
+        val accounts = getAvailableAccounts()
+        val initialAccounts = sessionRequestIdentityRequest.credentialStatements.map { statement ->
+            // Prefer session account
+            if (isValidIdentityForStatement(identities[sessionRequestAccount.identityId]!!, statement))
+                sessionRequestAccount
+            else {
+                // Otherwise find any account that can prove the statement
+                accounts.find {
+                    isValidIdentityForStatement(
+                        identities[it.identityId]!!,
+                        statement
+                    )
+                }
+            }
+        }
+
+        sessionRequestIdentityProofAccounts = if (initialAccounts.any { it == null }) {
+            // Check whether any statement does not have any account, which issuer is allowed.
+            val noCompatibleIssuers =
+                sessionRequestIdentityRequest.credentialStatements.any { statement ->
+                    accounts.none { account ->
+                        (statement.idQualifier as IdentityQualifier).issuers.contains(
+                            identities[account.identityId]!!.identityProviderId.toLong()
+                        )
+                    }
+                }
+            sessionRequestIdentityProofProvable = if (noCompatibleIssuers) {
+                ProofProvableState.NoCompatibleIssuer
+            } else {
+                ProofProvableState.UnProvable
+            }
+
+            initialAccounts.map { it ?: sessionRequestAccount }.toMutableList()
+        } else {
+            sessionRequestIdentityProofProvable = ProofProvableState.Provable
+            initialAccounts.requireNoNulls().toMutableList()
+        }
+        return true
+    }
+
     /**
      * Extracts the account address from a map of session namespaces. This works as the account
      * format follows the convention: `ccd:network:accountAddress`, which is defined by what we set
@@ -682,6 +883,13 @@ private constructor(
         when (method) {
             REQUEST_METHOD_SIGN_AND_SEND_TRANSACTION -> handleSignTransactionRequest(params)
             REQUEST_METHOD_SIGN_MESSAGE -> handleSignMessage(params)
+            REQUEST_METHOD_VERIFIABLE_PRESENTATION -> {
+                if (initializeProofRequestSession(params))
+                    mutableStateFlow.tryEmit(
+                        createIdentityProofRequestState(0)
+                    )
+            }
+
             else -> {
                 val unsupportedMethodMessage =
                     "Received an unsupported WalletConnect method request: $method"
@@ -900,6 +1108,28 @@ private constructor(
         }
     }
 
+    private suspend fun getAccountAttributes(
+        account: Account,
+        password: String
+    ): AttributeRandomness? {
+        val storageAccountDataEncrypted = account.encryptedAccountData
+        if (TextUtils.isEmpty(storageAccountDataEncrypted)) {
+            Log.e("account_has_no_encrypted_data")
+
+            mutableEventsFlow.tryEmit(
+                Event.ShowFloatingError(
+                    Error.CryptographyFailed
+                )
+            )
+            return null
+        }
+        val decryptedJson = App.appCore.getCurrentAuthenticationManager()
+            .decryptInBackground(password, storageAccountDataEncrypted)
+        val storageAccountData =
+            App.appCore.gson.fromJson(decryptedJson, StorageAccountData::class.java)
+        return storageAccountData.getAttributeRandomness()
+    }
+
     private fun decryptAccountAndApproveRequest(password: String) = viewModelScope.launch {
         val reviewState = checkNotNull(state as? State.SessionRequestReview) {
             "Approval is only possible in a session request review state"
@@ -918,10 +1148,11 @@ private constructor(
         }
         val decryptedJson = App.appCore.getCurrentAuthenticationManager()
             .decryptInBackground(password, storageAccountDataEncrypted)
+
         if (decryptedJson != null) {
-            val accountKeys =
+            val storageAccountData =
                 App.appCore.gson.fromJson(decryptedJson, StorageAccountData::class.java)
-                    .accountKeys
+            val accountKeys = storageAccountData.accountKeys
 
             when (reviewState) {
                 is State.SessionRequestReview.SignRequestReview ->
@@ -929,6 +1160,18 @@ private constructor(
 
                 is State.SessionRequestReview.TransactionRequestReview ->
                     signAndSubmitRequestedTransaction(accountKeys)
+
+                is State.SessionRequestReview.IdentityProofRequestReview -> {
+                    val attributeRandomness = HashMap<String, AttributeRandomness?>()
+                    attributeRandomness[storageAccountData.accountAddress] =
+                        storageAccountData.getAttributeRandomness()
+                    reviewState.chosenAccounts.forEach {
+                        if (!attributeRandomness.contains(it.address)) {
+                            attributeRandomness[it.address] = getAccountAttributes(it, password)
+                        }
+                    }
+                    createProofPresentation(reviewState.chosenAccounts, attributeRandomness)
+                }
             }
         } else {
             Log.e("failed_account_decryption")
@@ -942,6 +1185,136 @@ private constructor(
                 )
             )
         }
+    }
+
+    private fun createProofPresentation(
+        chosenAccounts: List<Account>,
+        randomnessMap: Map<String, AttributeRandomness?>,
+    ) {
+        val request = sessionRequestIdentityRequest
+
+        var accountIterator = chosenAccounts.iterator()
+        val commitmentInputs = try {
+            request.credentialStatements.map { statement ->
+                val account = accountIterator.next()
+                val attributeRandomness = randomnessMap[account.address]
+                val identity = getIdentity(account)
+
+                if (attributeRandomness == null) {
+                    Log.e("Attribute randomness is not available for account ${account.address}")
+                    mutableEventsFlow.tryEmit(
+                        Event.ShowFloatingError(
+                            Error.InternalError
+                        )
+                    )
+                    return
+                }
+                if (identity == null) {
+                    Log.e("Identity is not available for account ${account.address}")
+                    mutableEventsFlow.tryEmit(
+                        Event.ShowFloatingError(
+                            Error.InternalError
+                        )
+                    )
+                    return
+                }
+
+                val identityProviderIndex = identity.identityProviderId
+                val randomness: MutableMap<AttributeType, String> = mutableMapOf()
+                val attributeValues: MutableMap<AttributeType, String> = mutableMapOf()
+
+                statement.statement.forEach { stat ->
+                    val attributeTag = stat.attributeTag
+                    val attributeType = AttributeType.fromJSON(attributeTag)
+                    randomness[attributeType] = attributeRandomness.attributesRand[attributeTag]!!
+                    attributeValues[attributeType] =
+                        identity.identityObject!!.attributeList.chosenAttributes[attributeTag]!!
+                }
+
+                AccountCommitmentInput.builder()
+                    .issuer(UInt32.from(identityProviderIndex))
+                    .values(attributeValues)
+                    .randomness(randomness)
+                    .build()
+            }
+        } catch (e: Exception) {
+            Log.e("Failed to build commitment inputs", e)
+            mutableEventsFlow.tryEmit(
+                Event.ShowFloatingError(
+                    Error.InternalError
+                )
+            )
+            return
+        }
+
+        val network = if (BuildConfig.ENV_NAME == "production") Network.MAINNET else Network.TESTNET
+        accountIterator = chosenAccounts.iterator()
+        val qualifiedRequest = try {
+            request.qualify { stat ->
+                stat.qualify(
+                    CredentialRegistrationId.from(accountIterator.next().credential!!.getCredId()!!),
+                    network
+                )
+            }
+        } catch (e: Exception) {
+            Log.e("Failed to qualify request.", e)
+            mutableEventsFlow.tryEmit(
+                Event.ShowFloatingError(
+                    Error.InternalError
+                )
+            )
+            return
+        }
+
+        ProxyRepository().getIGlobalInfo({ globalInfo ->
+            val cryptographicParams: GlobalParams = globalInfo.value
+
+            val input = Web3IdProofInput.builder()
+                .request(qualifiedRequest)
+                .commitmentInputs(commitmentInputs)
+                .globalContext(
+                    CryptographicParameters.builder()
+                        .genesisString(cryptographicParams.genesisString)
+                        .bulletproofGenerators(BulletproofGenerators.from(cryptographicParams.bulletproofGenerators))
+                        .onChainCommitmentKey(PedersenCommitmentKey.from(cryptographicParams.onChainCommitmentKey))
+                        .build()
+                )
+                .build()
+
+            try {
+                val proof = Web3IdProof.getWeb3IdProof(input)
+
+                val wrappedProof = VerifiablePresentationWrapper(
+                    verifiablePresentationJson = proof
+                )
+                respondSuccess(result = Gson().toJson(wrappedProof))
+
+                mutableStateFlow.tryEmit(State.Idle)
+                handleNextOldestPendingSessionRequest()
+            } catch (e: Exception) {
+                Log.e("failed_creating_verifiable_presentation", e)
+                respondError(
+                    message = "Unable to create verifiable presentation: internal error"
+                )
+                mutableEventsFlow.tryEmit(
+                    Event.ShowFloatingError(
+                        Error.CryptographyFailed
+                    )
+                )
+            }
+        },
+            {
+                Log.e("Failed to retrieve global cryptographic parameters", it)
+                mutableEventsFlow.tryEmit(
+                    Event.ShowFloatingError(
+                        Error.LoadingFailed
+                    )
+                )
+            })
+    }
+
+    fun getIdentity(account: Account): Identity? {
+        return sessionRequestIdentityProofIdentities[account.identityId]
     }
 
     private suspend fun signRequestedMessage(accountKeys: AccountData) {
@@ -1064,7 +1437,14 @@ private constructor(
                 mutableStateFlow.tryEmit(State.Idle)
             }
 
-            is State.AccountSelection,
+            is State.AccountSelection -> {
+                if ((state as State.AccountSelection).forIdentityProof) {
+                    rejectSessionRequest()
+                } else {
+                    rejectSessionProposal()
+                }
+            }
+
             is State.SessionProposalReview -> {
                 rejectSessionProposal()
             }
@@ -1090,14 +1470,17 @@ private constructor(
         when (val state = this.state) {
             is State.AccountSelection -> {
                 Log.d(
-                    "switching_back_to_session_proposal_review"
+                    "switching_back_from_account_selection"
                 )
 
                 mutableStateFlow.tryEmit(
-                    State.SessionProposalReview(
-                        selectedAccount = state.selectedAccount,
-                        appMetadata = state.appMetadata,
-                    )
+                    if (state.forIdentityProof)
+                        createIdentityProofRequestState(state.identityProofPosition!!)
+                    else
+                        State.SessionProposalReview(
+                            selectedAccount = state.selectedAccount,
+                            appMetadata = state.appMetadata,
+                        )
                 )
                 true
             }
@@ -1234,8 +1617,11 @@ private constructor(
         class AccountSelection(
             val selectedAccount: Account,
             val accounts: List<Account>,
+            val identityProofPosition: Int?,
             val appMetadata: AppMetadata,
-        ) : State
+        ) : State {
+            val forIdentityProof = identityProofPosition != null
+        }
 
         /**
          * Explicitly waiting for a session request after receiving a request URI.
@@ -1273,6 +1659,20 @@ private constructor(
                 account = account,
                 appMetadata = appMetadata,
                 canApprove = true,
+            )
+
+            class IdentityProofRequestReview(
+                val request: UnqualifiedRequest,
+                val chosenAccounts: List<Account>,
+                val currentStatement: Int,
+                val provable: ProofProvableState,
+                connectedAccount: Account,
+                appMetadata: AppMetadata,
+            ) : SessionRequestReview(
+                account = connectedAccount,
+                appMetadata = appMetadata,
+                // Check that we can prove the statement with current accounts
+                canApprove = provable == ProofProvableState.Provable
             )
         }
 
@@ -1339,6 +1739,16 @@ private constructor(
          * There is a problem decrypting the account or the crypto library failed.
          */
         object CryptographyFailed : Error
+
+        /**
+         * An internal error occurred.
+         */
+        object InternalError : Error
+
+        /**
+         * The functionality is only supported by a seed phrase based wallet.
+         */
+        object NotSeedPhraseWalletError : Error
     }
 
     sealed interface Event {
@@ -1356,6 +1766,23 @@ private constructor(
         ) : Event
     }
 
+    /**
+     * Wrapper for sending the verifiable presentation as JSON over WalletConnect. This is
+     * required to prevent WalletConnect from automatically parsing the JSON as an object
+     * on the dApp side.
+     */
+    private class VerifiablePresentationWrapper(
+        val verifiablePresentationJson: String
+    )
+
+    /**
+     * Wrapper for receiving parameters as JSON over WalletConnect. This is required to allow a
+     * dApp to send bigint values from the Javascript side.
+     */
+    private class WalletConnectParamsWrapper(
+        val paramsJson: String
+    )
+
     private companion object {
         private const val WC_URI_PREFIX = "wc:"
         private const val WC_URI_REQUEST_ID_PARAM = "requestId"
@@ -1364,5 +1791,6 @@ private constructor(
 
         private const val REQUEST_METHOD_SIGN_AND_SEND_TRANSACTION = "sign_and_send_transaction"
         private const val REQUEST_METHOD_SIGN_MESSAGE = "sign_message"
+        private const val REQUEST_METHOD_VERIFIABLE_PRESENTATION = "request_verifiable_presentation"
     }
 }