+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Setting up and updating validator pool commission rates
 - Support of WalletConnect CCD transfer requests
+- Support for WalletConnect verifiable presentation requests (for identity proofs)
 
 ### Removed
 - Revealing identity attributes when creating an account
@@ -22,6 +23,7 @@
 - An issue where the identity name was off-center when the edit name icon was visible
 - An issue where exporting transaction logs for an account without any transactions would be stuck at 0%
 - "Invalid WalletConnect request" message repeatedly shown if received a request with unsupported transaction type
+- Exported private key for file-based initial account being incompatible with concordium-client
 
 ### Changed
 - Suggest running a recovery when facing account or identity creation errors